+++ src/components/SendForm/SendForm.tsx
@@ -4,7 +4,10 @@ import cn from "classnames";
 import BN from "bn.js";
 import axios from "axios";
 import * as Sentry from "@sentry/browser";
+import { fetchBalance } from "@wagmi/core";
+import { useAccount, useContract, useNetwork, useSigner, useSwitchNetwork } from "wagmi";
 
+import Web3 from "web3";
 import {
     changeNetworkAtMetamask,
     NetworkType,
@@ -13,14 +16,14 @@ import {
     networkNames,
     networkToId,
     getAddressLink,
+    NETWORK,
 } from "../../utils/network";
-import { ensToAddress, isAddress } from "../../utils/wallet";
+import { getShortAddress } from "../../utils/wallet";
 import {
     beautifyTokenBalance,
     CUSTOM_TOKENS,
     fromHRToBN,
     getCovalentUrl,
-    getTokenContractFactory,
     nativeTokensAddresses,
     toHRNumberFloat,
 } from "../../utils/tokens";
@@ -43,6 +46,8 @@ import { PageNotFound } from "../PageNotFound";
 import { useCurrentTokenData } from "../../hooks/useCurrentTokenData";
 import { GradientHref } from "../../ui-kit/components/GradientHref";
 import GenUrlPopup from "../GenUrlPopup";
+import CONTRACT_ERC20 from "../../contracts/ERC20.json";
+import { useEns } from "../../hooks/useEns";
 
 import "./SendForm.scss";
 
@@ -62,8 +67,12 @@ interface ApproveFormProps {
 }
 
 const SendForm = ({ onConnect }: ApproveFormProps) => {
-    const { address, chainId, web3, newCustomToken } = useContext(StateContext);
-    const networkName = chainId ? idToNetwork[chainId] : undefined;
+    const { data: signer } = useSigner();
+    const { address } = useAccount();
+    const { chain } = useNetwork();
+    const { switchNetwork } = useSwitchNetwork();
+    const networkName = chain?.id ? idToNetwork[chain?.id] : undefined;
+    const { newCustomToken } = useContext(StateContext);
     const [toAddress, setToAddress] = useState<string>();
     const [value, setValue] = useState<number>();
     const [selectedToken, setSelectedToken] = useState<string>(); // address
@@ -75,12 +84,23 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
     const [balances, setBalances] = useState<BalanceType[]>([]);
     const [genUrl, setGenUrl] = useState<string>();
     const [allowance, setAllowance] = useState<string>();
+    const { isEnsAddress, isEnsName, ensName, ensAddress, isEnsNameLoading, isAvvyNameLoading, isAvvyName, avvyName } =
+        useEns(toAddress);
 
     const { requestTokenData, requestToken } = useRequestToken();
     const isDisabledByToken = requestTokenData && networkName !== requestTokenData.networkName;
     const hasRequestToken = !!requestTokenData;
 
     const currentToken = useCurrentTokenData(balances, selectedToken, requestTokenData);
+    const currentTokenContract = useContract({
+        address: currentToken?.token_address,
+        abi: CONTRACT_ERC20,
+        signerOrProvider: signer,
+    });
+
+    useEffect(() => {
+        trackEvent("openSendPage", { address });
+    }, [address]);
 
     useEffect(() => {
         if (hasRequestToken && currentToken) {
@@ -97,16 +117,13 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
         }
     }, [requestToken]);
 
-    const isCorrectData = isAddress(toAddress) && (value ?? 0) > 0 && selectedToken;
     const currentTokenBalance = currentToken
         ? toHRNumberFloat(new BN(currentToken.balance), +currentToken.decimals)
         : 0;
     const hasAllowance = !!(allowance && allowance !== "0");
 
-    const getTokenContract = getTokenContractFactory(web3);
-
     const onMount = async () => {
-        if (chainId && address && web3 && !requestTokenData?.token) {
+        if (chain && address && !requestTokenData?.token) {
             let longRequestTimeoutId;
             try {
                 setIsBalanceRequestLoading(true);
@@ -114,7 +131,7 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
                     Sentry.captureMessage("Long Covalent request");
                 }, 10000);
 
-                const response = await axios.get(getCovalentUrl(chainId, address));
+                const response = await axios.get(getCovalentUrl(chain.id, address));
                 clearTimeout(longRequestTimeoutId);
                 setIsBalanceRequestLoading(false);
 
@@ -149,20 +166,16 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
                 })),
             ];
             for (const token of tokens) {
-                const tokenContract = getTokenContract(token.address);
-                if (!tokenContract) {
-                    return;
-                }
-
                 try {
-                    const balance = await tokenContract.methods.balanceOf(address.toLowerCase()).call();
+                    // @ts-ignore
+                    const balance = await fetchBalance({ address, token: token.address });
 
                     const newTokenBalance = {
                         token_address: token.address,
                         name: token.id,
                         symbol: token.symbol,
                         decimals: token.decimals,
-                        balance,
+                        balance: balance.value.toString(),
                     } as BalanceType;
                     setBalances((oldBalances) => {
                         const newBalances = arrayUniqueByKey(
@@ -201,11 +214,21 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
 
     useEffect(() => {
         onMount();
-    }, [chainId, address, web3, requestToken]);
+    }, [networkName, address, requestToken]);
 
-    const handleNetworkChange = useCallback((event) => {
-        changeNetworkAtMetamask(event.target.value);
-    }, []);
+    const handleNetworkChange = useCallback(
+        (event) => {
+            // @ts-ignore
+            const chainId = networkToId[event.target.value];
+            const hexChainId = Web3.utils.toHex(chainId);
+            if (switchNetwork && hexChainId) {
+                // @ts-ignore
+                switchNetwork(hexChainId);
+                trackEvent("NetworkChanged", { address, network: event.target.value });
+            }
+        },
+        [switchNetwork, address]
+    );
 
     const handleAddressChange = (event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
         setToAddress(event.target.value);
@@ -213,6 +236,7 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
 
     const handleTokenChange = (event: SelectChangeEvent) => {
         setSelectedToken(event.target.value);
+        trackEvent("TokenChanged", { address, token: event.target.value });
     };
 
     const handleValueChange = (event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
@@ -239,33 +263,30 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
     };
 
     const cancelApprove = async () => {
-        const tokenContract = getTokenContract(currentToken?.token_address ?? "");
-        if (!tokenContract || !web3 || !networkName) {
+        if (!networkName || !currentTokenContract) {
             addErrorNotification("Error", "No network");
             return;
         }
 
         try {
             setIsCancelApproveLoading(true);
-            await tokenContract.methods
-                .approve(await ensToAddress(networkName, toAddress), "0")
-                .send({ from: address, maxPriorityFeePerGas: null, maxFeePerGas: null })
-                .on("transactionHash", async (txHash: string) => {
-                    const symbol = await tokenContract.methods.symbol().call();
+            const tx = await currentTokenContract.approve(ensAddress, "0");
+            const symbol = await currentTokenContract.symbol();
+            await tx.wait();
 
-                    // eslint-disable-next-line max-len
-                    // Disclaimer: since all data above are always public on blockchain, so there’s no compromise of privacy. Beware however, that underlying infrastructure on users, such as wallets or Infura might log sensitive data, such as IP addresses, device fingerprint and others.
-                    trackEvent("APPROVED_REVOKE_SENT", {
-                        location: window.location.href,
-                        source: "Send Page",
-                        chainId: networkToId[networkName],
-                        txHash,
-                        fromAddress: address,
-                        toAddress,
-                        tokenAddress: currentToken?.token_address,
-                        tokenSymbol: symbol,
-                    });
-                });
+            addSuccessNotification("Success", "Cancel allowance completed");
+            // eslint-disable-next-line max-len
+            // Disclaimer: since all data above are always public on blockchain, so there’s no compromise of privacy. Beware however, that underlying infrastructure on users, such as wallets or Infura might log sensitive data, such as IP addresses, device fingerprint and others.
+            trackEvent("APPROVED_REVOKE_SENT", {
+                location: window.location.href,
+                source: "Send Page",
+                chainId: networkToId[networkName],
+                txHash: tx.hash,
+                fromAddress: address,
+                toAddress,
+                tokenAddress: currentToken?.token_address,
+                tokenSymbol: symbol,
+            });
             setAllowance(undefined);
         } catch (error) {
             // @ts-ignore
@@ -285,44 +306,39 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
     };
 
     const handleApprove = async () => {
-        if (currentToken && value && toAddress && address && web3 && networkName) {
+        if (currentToken && currentTokenContract && value && toAddress && address && networkName) {
             setGenUrl(undefined);
             setTrxHash("");
             setTrxLink("");
             setIsApproveLoading(true);
 
             const valueBN = fromHRToBN(value, +currentToken.decimals).toString();
-            const tokenContract = getTokenContract(currentToken.token_address);
-            const ensAddress = await ensToAddress(networkName, toAddress);
             try {
-                await tokenContract?.methods
-                    .approve(ensAddress, valueBN)
-                    .send({ from: address, maxPriorityFeePerGas: null, maxFeePerGas: null })
-                    .on("transactionHash", (hash: string) => {
-                        setTrxHash(hash);
-                        setTrxLink(getTrxHashLink(hash, networkName));
-                        setGenUrl(
-                            generateUrl({
-                                address: currentToken?.token_address,
-                                from: address ?? "",
-                                to: toAddress ?? "",
-                                value: fromHRToBN(value ?? 0, +currentToken.decimals).toString(),
-                                chain: networkName,
-                            })
-                        );
-                        // eslint-disable-next-line max-len
-                        // Disclaimer: since all data above are always public on blockchain, so there’s no compromise of privacy. Beware however, that underlying infrastructure on users, such as wallets or Infura might log sensitive data, such as IP addresses, device fingerprint and others.
-                        trackEvent("APPROVE_SENT", {
-                            location: window.location.href,
-                            chainId: networkToId[networkName],
-                            txHash: hash,
-                            fromAddress: address,
-                            toAddress,
-                            tokenAddress: currentToken?.token_address,
-                            tokenSymbol: getTokenName(selectedToken),
-                            tokenAmount: value,
-                        });
-                    });
+                const tx = await currentTokenContract.approve(ensAddress, valueBN);
+                setTrxHash(tx.hash);
+                setTrxLink(getTrxHashLink(tx.hash, networkName));
+                setGenUrl(
+                    generateUrl({
+                        address: currentToken?.token_address,
+                        from: address ?? "",
+                        to: ensAddress ?? "",
+                        value: fromHRToBN(value ?? 0, +currentToken.decimals).toString(),
+                        chain: networkName,
+                    })
+                );
+                // eslint-disable-next-line max-len
+                // Disclaimer: since all data above are always public on blockchain, so there’s no compromise of privacy. Beware however, that underlying infrastructure on users, such as wallets or Infura might log sensitive data, such as IP addresses, device fingerprint and others.
+                trackEvent("APPROVE_SENT", {
+                    location: window.location.href,
+                    chainId: networkToId[networkName],
+                    txHash: tx.hash,
+                    fromAddress: address,
+                    toAddress: ensAddress,
+                    tokenAddress: currentToken?.token_address,
+                    tokenSymbol: getTokenName(selectedToken),
+                    tokenAmount: value,
+                });
+                await tx.wait();
                 onSuccessApprove();
             } catch (error) {
                 // @ts-ignore
@@ -348,14 +364,9 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
     };
 
     const setAllowanceAsync = async () => {
-        if (isCorrectData && currentToken && networkName) {
-            const tokenContract = getTokenContract(currentToken.token_address);
-            if (tokenContract) {
-                const allowanceFromContract = await tokenContract.methods
-                    .allowance(address, await ensToAddress(networkName, toAddress))
-                    .call();
-                setAllowance(allowanceFromContract.toString());
-            }
+        if (address && currentToken && currentTokenContract && ensAddress) {
+            const allowanceFromContract = await currentTokenContract.allowance(address, ensAddress);
+            setAllowance(allowanceFromContract.toString());
         } else {
             setAllowance(undefined);
         }
@@ -363,7 +374,7 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
 
     useEffect(() => {
         setAllowanceAsync();
-    }, [isCorrectData]);
+    }, [address, ensAddress, currentToken]);
 
     if (requestToken && !requestTokenData) {
         return <PageNotFound />;
@@ -414,7 +425,16 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
                             </Select>
                         </FormControl>
 
-                        <div className="send-form__label">Recipient address</div>
+                        <div className="send-form__label">
+                            Recipient address
+                            <span className="send-form__label-additional">
+                                {(isEnsName || isAvvyName) && ensAddress && ` (${getShortAddress(ensAddress)})`}
+                            </span>
+                            <span className="send-form__label-additional">
+                                {isEnsAddress && ensName && ` (${ensName})`}
+                                {isEnsAddress && avvyName && ` (${avvyName})`}
+                            </span>
+                        </div>
                         <TextField
                             value={toAddress}
                             disabled={hasRequestToken}
@@ -514,7 +534,7 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
                             <Button
                                 onClick={cancelApprove}
                                 className="send-form__button"
-                                disabled={isCancelApproveLoading}
+                                disabled={isCancelApproveLoading || !ensAddress}
                             >
                                 {isCancelApproveLoading ? "Loading..." : "Cancel Approve"}
                             </Button>
@@ -525,7 +545,14 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
                         <Button
                             onClick={handleApprove}
                             className="send-form__button"
-                            disabled={!isCorrectData || isApproveLoading || hasAllowance || isDisabledByToken}
+                            disabled={
+                                !value ||
+                                !selectedToken ||
+                                !ensAddress ||
+                                isApproveLoading ||
+                                hasAllowance ||
+                                isDisabledByToken
+                            }
                         >
                             {isApproveLoading ? "Loading..." : "Approve"}
                         </Button>
@@ -536,6 +563,17 @@ const SendForm = ({ onConnect }: ApproveFormProps) => {
                     )}
                     {requestTokenData && requestTokenData.networkName !== networkName && (
                         <div className="send-form__error">Please change network to {requestTokenData.networkName}</div>
+                    )}
+                    {toAddress && !isEnsAddress && !isEnsName && !isAvvyName && (
+                        <div className="send-form__error">Please write correct recipient address</div>
+                    )}
+                    {isEnsNameLoading && <div className="send-form__warning">ENS resolve in progress</div>}
+                    {isAvvyNameLoading && <div className="send-form__warning">AVAX resolve in progress</div>}
+                    {isEnsName && !ensAddress && !isEnsNameLoading && (
+                        <div className="send-form__error">Can't resolve ENS address</div>
+                    )}
+                    {isAvvyName && !ensAddress && !isAvvyNameLoading && (
+                        <div className="send-form__error">Can't resolve AVAX address</div>
                     )}
                 </div>
             </div>